+++ lib/controller/categories_wise_quiz_list_page_controller2.dart
@@ -1,0 +1,32 @@
+
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:get/get.dart';
+import 'package:get_storage/get_storage.dart';
+
+import '../../static/Colors.dart';
+import '../../view/common/toast.dart';
+
+class CategoriesWiseQuizListPage extends GetxController {
+
+
+  var isDrawerOpen = false.obs;
+
+
+  void ref(){
+    onInit();
+  }
+  @override
+  void onInit() {
+
+
+    super.onInit();
+
+  }
+
+
+
+}
