+++ lib/controller/indevidual_quiz_about_page_controller.dart
@@ -1,0 +1,37 @@
+
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:get/get.dart';
+import 'package:get_storage/get_storage.dart';
+
+import '../../static/Colors.dart';
+import '../../view/common/toast.dart';
+
+class IndividualQuizAboutPagePage extends GetxController {
+
+
+  var aboutQuizText = "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been "
+      "the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type "
+      "and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into"
+      " electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of "
+      "Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus "
+      "PageMaker including versions of Lorem IpsumLorem Ipsum is simply dummy text of the printing and typesetting "
+      "industry. ".obs;
+
+  void ref(){
+    onInit();
+  }
+  @override
+  void onInit() {
+
+
+    super.onInit();
+
+  }
+
+
+
+}
