+++ lib/view/home_page/dash_board_page.dart
@@ -8,8 +8,8 @@ import '../../controller/dash_board_page_controller.dart';
 import '../../static/Colors.dart';
 
 import '../custom_drawer.dart';
+import 'quiz_categories_page.dart';
 import 'home_page.dart';
-import 'home_page2.dart';
 import 'my_profile_page.dart';
 
 class DashBoardPageScreen extends StatelessWidget {
@@ -36,13 +36,13 @@ class DashBoardPageScreen extends StatelessWidget {
         onTap: (int index){
           dashBoardPageController.selectedTabIndex(index);
           if(index==0){
-            dashBoardPageController.updateSelectedPage([Homepage2Screen()]);
+            dashBoardPageController.updateSelectedPage([HomepageScreen()]);
             // selectedPage(HomePage( ));
             return;
           }
 
           if(index==1){
-            dashBoardPageController.updateSelectedPage([HomepageScreen()]);
+            dashBoardPageController.updateSelectedPage([QuizCategoriesScreen()]);
             //  selectedPage(HomePage( ));
             // selectedPage= ShopPage( );
             return;
